+++ Source/Classes/Services/JSNotificationCenter.m
@@ -41,19 +41,19 @@ NSString *const CKTNotificationUserUpdated = @"userUpdated";
 // Notificaiton to WebRTC to mute / unmute the speaker
 NSString *const NOTIFICATION_AUDIO_SESSION_MUTE_SPEAKER = @"circuitkit.notification.AUDIO_SESSION_MUTE_SPEAKER";
 // Notificaiton from WebRTC that Video receive has been lost
-NSString *const NOTIFICATION_VIDEO_RECEIVE_STREAM_LOST =
-    @"circuitkit.notification.VIDEO_RECEIVE_STREAM_LOST";
+NSString *const NOTIFICATION_VIDEO_RECEIVE_STREAM_LOST = @"circuitkit.notification.VIDEO_RECEIVE_STREAM_LOST";
 // WebRTC failed to setup AVAudioSession during call
 NSString *const NOTIFICATION_AUDIO_SESSION_SETUP_FAILED = @"circuitkit.notification.AUDIO_SESSION_SETUP_FAILED";
 NSString *const NOTIFICATION_AUDIO_SESSION_CHANGE_CATEGORY_REQUESTED =
-@"circuitkit.notification.NOTIFICATION_AUDIO_SESSION_CHANGE_CATEGORY_REQUESTED";
+    @"circuitkit.notification.NOTIFICATION_AUDIO_SESSION_CHANGE_CATEGORY_REQUESTED";
 
 NSString *const CKTKeyEmpty = @"circuitkit.key.EMPTY";
 
 // KEY_EVENT_DATA is used in the WebRTC library libWebRTC.a, we must use that same naming convention
 NSString *const KEY_EVENT_DATA = @"circuitkit.key.DATA";
 NSString *const KEY_AUDIO_ACTIVATE = @"circuitkit.key.AUDIO_ACTIVATE";
-
+NSString *const KEY_CALL = @"circuitkit.key.CALL";
+NSString *const KEY_CALL_REPLACED = @"circuitkit.key.REPLACED_CALL_FLAG";
 
 static NSString *LOG_TAG = @"[JSNotificationCenter]";