+++ Source/Classes/Services/PubSubService.m
@@ -19,6 +19,7 @@
 //
 //
 
+#import "CKTService.h"
 #import "JSEngine.h"
 #import "JSNotificationCenter.h"
 #import "JSValue+an.h"
@@ -29,6 +30,7 @@
 
 typedef void (^EventCallbackZeroArg)(JSValue *value);
 typedef void (^EventCallbackOneArg)(JSValue *value);
+typedef void (^EventCallbackTwoArgs)(JSValue *value1, JSValue *value2);
 
 @interface PubSubService ()
 
@@ -84,8 +86,7 @@ static NSString *LOG_TAG = @"[PubSubService]";
     LOGI(LOG_TAG, @"processReceivedEvent (one argument) - received event (%@)", notificationEvent);
 
     // This method should only be called on the js thread, assert if otherwise.
-    NSAssert([NSThread currentThread] == [JSEngine sharedInstance].jsThread,
-             @"This method should only be called on the js thread!");
+    [self validateCurrentThread];
 
     @try {
         NSDictionary *pubDict = @{ CKTKeyEmpty : @"" };
@@ -98,6 +99,9 @@ static NSString *LOG_TAG = @"[PubSubService]";
 
         switch (event) {
             case JSEventBasicSearchResults:
+            case JSEventCallEnded:
+            case JSEventCallStatus:
+            case JSEventCallIncoming:
             case JSEventConnectionStateChanged:
             case JSEventConversationCreated:
             case JSEventConversationUpdated:
@@ -110,14 +114,15 @@ static NSString *LOG_TAG = @"[PubSubService]";
             case JSEventUserSettingsChanged:
             case JSEventUserUpdated: {
                 pubDict = @{KEY_EVENT_DATA : data};
+                break;
             }
-
             default: {
                 LOGD(LOG_TAG, @"Unsupported event: %@", notificationEvent);
             }
-
-                [self processEvent:event withData:data andNotification:notificationEvent];
         }
+
+        // Send the event to the other handlers and then send the notification
+        [self processEvent:event withData:data andNotification:notificationEvent];
     }
     @catch (NSException *exception)
     {
@@ -126,6 +131,20 @@ static NSString *LOG_TAG = @"[PubSubService]";
 }
 
 #pragma mark - Internal methods
+
+// Used when processing received events to validate the current thread is the jsThread
+- (void)validateCurrentThread
+{
+    if ([JSEngine sharedInstance].jsThread && ![NSThread currentThread].isCancelled) {
+        NSAssert([NSThread currentThread] == [JSEngine sharedInstance].jsThread,
+                 @"This method should only be called on the js thread!");
+    } else {
+        // This is done in case we receive an event during stop/start the JSEngine e.g., jsThread is nil but JSRunLoop
+        // has not yet stopped or jsThread is present but currentThread was the previously cancelled thread.
+        NSAssert([[NSThread currentThread].name isEqualToString:kJSRunloopName],
+                 @"This method should only be called on the (%@)", kJSRunloopName);
+    }
+}
 
 - (void)subscribe:(JSEvent)event
 {
@@ -153,11 +172,51 @@ static NSString *LOG_TAG = @"[PubSubService]";
     }
 }
 
+- (void)processReceivedEvent:(JSEvent)event data1:(JSValue *)value1 data2:(JSValue *)value2
+{
+    NSString *notificationEvent = [self topicStringFromPublishedEvent:event];
+    LOGI(LOG_TAG, @"processReceivedEvent (two arguments) - received event (%@)", notificationEvent);
+
+    // This method should only be called on the js thread, assert if otherwise.
+    [self validateCurrentThread];
+
+    @try {
+        id data1 = [value1 an_dataFromJSValue];
+        id data2 = [value2 an_dataFromJSValue];
+
+        if (data1 == nil) {
+            LOGE(LOG_TAG, @"Event %@ is supposed to have at least one piece data, but doesn't", notificationEvent);
+            return;
+        }
+
+        NSDictionary *pubDict;
+
+        // This section handles events that have an optional second piece of data
+        if (event == JSEventCallEnded) {
+            // Check if this is a replaced call (a 1-2-1 call changed to a group call)
+            NSNumber *callReplaced = @NO;
+            if (data2) {
+                callReplaced = (NSNumber *)data2;
+            }
+            pubDict = @{KEY_CALL : data1, KEY_CALL_REPLACED : callReplaced};
+        }
+        // Send the event to the other handlers and then send the notification
+        [self processEvent:event withData:pubDict andNotification:notificationEvent];
+    }
+    @catch (NSException *exception)
+    {
+        JS_SERVICE_LOG_EXCEPTION(exception);
+    }
+}
+
 - (void)initEventCallbacks
 {
     self.eventCallbacks = @[
         [[self zeroArgBlockForEvent:JSEventUnknown] copy],
         [[self oneArgBlockForEvent:JSEventBasicSearchResults] copy],
+        [[self oneArgBlockForEvent:JSEventCallEnded] copy],
+        [[self oneArgBlockForEvent:JSEventCallIncoming] copy],
+        [[self oneArgBlockForEvent:JSEventCallStatus] copy],
         [[self oneArgBlockForEvent:JSEventConnectionStateChanged] copy],
         [[self oneArgBlockForEvent:JSEventConversationCreated] copy],
         [[self oneArgBlockForEvent:JSEventConversationUpdated] copy],
@@ -180,6 +239,9 @@ static NSString *LOG_TAG = @"[PubSubService]";
         topics = @[
             @"UNKNOWN",
             CKTNotificationBasicSearchResults,
+            CKTNotificationCallEnded,
+            CKTNotificationCallIncoming,
+            CKTNotificationCallStatus,
             CKTNotificationConnectionStateChange,
             CKTNotificationConversationCreated,
             CKTNotificationConversationUpdated,
@@ -214,4 +276,8 @@ static NSString *LOG_TAG = @"[PubSubService]";
     return ^(JSValue *value) { [self processReceivedEvent:event data:value]; };
 }
 
+- (EventCallbackTwoArgs)twoArgBlockForEvent:(JSEvent)event
+{
+    return ^(JSValue *value1, JSValue *value2) { [self processReceivedEvent:event data1:value1 data2:value2]; };
+}
 @end