+++ Source/scripts/sdkInterface.js
@@ -2,3 +2,9 @@
 /*exported sdkClient*/
 
 var sdkClient = new Circuit.Client();
+
+
+//---------------------------------------------------------------------------
+// Inject Promise in non-angular modules
+//---------------------------------------------------------------------------
+Circuit.CallStatsHandler.overridePromise(Promise);